SecurityPkg: Add a PCD to skip Tcg2Smm ACPI table measurement
The current measurements are being made against 1) a compiled ASL file
before runtime-specific values are patched and 2) a C structure for
an ACPI table header (before runtime-specific values are patched). These
measurements are of dubious value to the OS or any remote attestation
service, and are difficult to pre-calculate at build time.

This change enables a platform to opt-out of these measurements.

We will consider adding them back when we have time to take a harder
look at how they are implemented and how they could be made more useful
to any higher-level consumers.

Signed-off-by: Bret Barkelew <bret.barkelew@microsoft.com>

cherry-pick from 3e855e5d95