SecurityPkg: Remove custom mode setting during PK deletion
This change will remove the side effect of automatically setting the
setup mode to `CustomMode` after deleting PK. The usage of `CustomMode`
is largely avoided in the scope of Project Mu to reduce attack surfaces.
This change is to align with the original behavior provided through
`MsSecureBootLib`.

- [x] Impacts functionality?
- [x] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

Tested on Q35 virtual platform.

N/A