BaseTools/Plugin: Add Rust Environment Check build plugin
Background

Firmware developer's machines are often not setup for Rust. As more
firmware developers begin working with Rust for the first time, this
plugin is used to provide early and direct feedback about the
developer's environment so it can successfully build Rust code using
the tools commonly used in the firmware build process.

The plugin is run when the `rust-ci` scope is specified which is used
by build scripts to opt into Rust plugin support.

The entire plugin takes ~1.4 sec to run on average so build time is
not meaningfully impacted.

---

Configuration

A feature that the plugin provides is the ability to set certain
tool exclusions. Using this, a build wrapper sets an environment
variable (chosen as the input method for its simplicity) to exclude a
list of tools.

For example, a GitHub workflow that only runs CodeQL for C code,
might need cargo installed for Rust compilation as part of the
package build, but might not need cargo tarpaulin for code coverage.

The GitHub workflow can set the environment variable on the build
step to opt out of verifying those tools.

This exclusion option is not intended to be used often as most local
developers and build environments are expected to have the base set
of tools needed to build and test the code when the plugins scope is
present.

It is very simple to set the environment variable in a CI environment
like an Azure Pipeline step as shown below for a step that builds
from information in a matrix with CodeQL enabled.

```
    - name: CI Build
      env:
        RUST_ENV_CHECK_TOOL_EXCLUSIONS: "cargo fmt, cargo tarpaulin"
        STUART_CODEQL_PATH:
		  ${{ steps.cache_key_gen.outputs.codeql_cli_ext_dep_dir }}
      run: stuart_ci_build -c .pytool/CISettings.py -t DEBUG -p \
	       ${{ matrix.package }} -a ${{ matrix.archs }} \
		   TOOL_CHAIN_TAG=${{ matrix.tool_chain_tag }} --codeql
```

Another feature the plugin provides is the ability to enforce version
requirements for some Rust related tools needed for build. Repository
owners can optionally require specific versions of tools to be
installed. This is helpful because it pins the supported toolset at a
given point in the repos timeline. As an example, if a repository is
on Rust 1.73 (as specified in the rust-toolchain.toml file), "cargo
install" on the latest tools might not work as they might require
Rust 1.76. With this check, the user can simply be information to
install the tool version known to be compatible at that time in the
repo rather than receive ambiguous compilation failures when trying
to install the latest tools.

These tool version can be specified in the rust-toolchain.toml file
at the root of the repo. The following example updates the install
command suggestion output from the plugin to include the specific
versions given for cargo-make and cargo-tarpaulin.

```
[toolchain]
channel = "1.73.0"

[tool]
cargo-tarpaulin = "0.27.3"
cargo-make = "0.37.9"
```

Co-authored-by: Joey Vagedes <joey.vagedes@gmail.com>
Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>