SecurityPkg: Add an assert to TCG log function if log is full (#257)
Currently, if the TCG log fills up, the firmware will boot only logging
some errors and the OS may or may not fail depending on scenario and
configuration. This PR adds an assert so that these truncations can be
found in testing rather then having to wait for failures in production.

- [X] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

N/A

N/A

cherry-pick from 1c115e6083

Signed-off-by:  Chris Fernald <chfernal@microsoft.com>