MdePkg: Add EnableInterruptsAndSleep Function
This patch adds support for IA32 and X64 platforms to enable
interrupts and sleep.

[SQUASH ON REBASE] Clang/Gcc: Move EnableInterruptsAndSleep out of EnableInterrupts.nasm. (#1712)

EnableInterrupts.nasm does not build for clang/gcc, so they cannot link
to EnableInterruptsAndSleep.

Squash with 4969b55336aa45859a8f594fca5c175506bc81a8.

- [ ] Impacts functionality? No.
- [ ] Impacts security? No.
- [ ] Breaking change? No.
- [ ] Includes tests? No.
- [ ] Includes documentation? No.

No. But the opposite. I cannot build HyperV-UEFI with Clang or Gcc
because of this.

Easy.