MdeModulePkg: Support Additional Event TPLs
This patch adds more support to the event subsystem by allowing
a platform to conditionally install a protocol that provides an
interface to the internal WaitForEvent() that does not have
TPL requirements. This protocol removes the TPL_APPLICATION requirement
in CoreWaitForEvent and makes the event TPL volatile as it can change
in the situation where the CPU has been instructed to sleep until an
interrupt arrives.