BaseTools/Plugin/RustEnvironmentCheck: Use pytools Rust helpers (#1038)
The plugin implementation has moved to edk2-pytool-extensions so it can
be reused for plugins targeting different scenarios such as
public/generic (this plugin) or custom internal environments that may
need to add on additional functionality.

This simplifies this plugin's implementation significantly.

---

Cherry pick from release/202311

---

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

- Verified plugin still detects errors properly
- Unit tests added in edk2-pytool-extensions

An `id` has been added to the plugin YAML file (`rust-env-check`). This
retains the same scope as before (`rust-ci`) but allows a custom version
of the plugin to override this version by specifying:

- `"id_override": "rust-env-check"`

In its YAML file. Otherwise, no integration work is needed.

> There is an example of code that sets `id_override` (via generated
YAML)

[here](https://github.com/microsoft/mu_basecore/blob/7cf3382ebaf5b5b5a2487d7d13e910c1c7dac233/BaseTools/Edk2ToolsBuild.py#L106-L107)
for reference.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>