MdeModulePkg, StandaloneMmPkg: Add resource descriptor hob v2 for memory attributes usage (#1258)
This change adds the implementation of supporting "resource descriptor
hob type 2", which will extends the existing definition of resource
descriptor hob with an attribute field.

This field can be used to provide the actual attributes needs to be set
for the memory region, instead of relying on the state of the system.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?
- [ ] Backport to release branch?

This change was tested on QEMU Q35 and SBSA systems and booted to
Windows.

For platforms needing to propagate specific memory attributes
(especially the cache attributes) to the DXE phase,
`BuildResourceDescriptorWithCacheHob` should be used to produce resource
descriptors (v2).