Update memory map descriptor merging logic to honor the bucket type information (#1429)
The current memory map merging logic in `CoreAddRange` and
`MergeMemoryMapByAttribute` only checks the memory type without
considering the bucket type. However, the `CoreGetMemoryMap` routine
will only report the entries that are entirely inside a specific bucket
type. This will confuse the consumer and report false free memory that
belongs to special memory buckets.

This change fixes the issue by gating the merging logic by checking the
bucket type first. The merging is only allowed either if the bucket type
is not one of the special types, or if the bucket type is the same for
both entries.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?
- [x] Backport to release branch?

This was tested on proprietary hardware.

N/A

---------

Signed-off-by: kuqin12 <42554914+kuqin12@users.noreply.github.com>
Co-authored-by: Michael Kubacki <michael.kubacki@microsoft.com>