MdeModulePkg: Fix CoherentPciIoUnmap (#1454)
## Description

Fixes CoherentPciIoUnmap and properly undoes CoherentPciIoMap
Add check for above 4GB case so that we only copy the data and free it
for the bounce buffer case.
Currently, it will try to copy and free no matter what and crash.

When using CoherentPciIo functionality in NonDiscoverablePciDeviceIo,
fixes error seen with connected devices.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?
- [x] Backport to release branch?

## How This Was Tested

Tested on physical arm platform

## Integration Instructions

N/A