StandaloneMmPkg: MmCommBufferUpdate: Adding new protocol
The current Standalone MM model does not coalesce the runtime memory when
the unblock memory request in PEI allocate runtime memory.

This will lead to fragmented memory map and potentially hibernation
failure.

This change adds a protocol definition that allows an event to be
signaled when all the PEI-allocated runtime memory migrate to permanent
runtime memory.

Signed-off-by: Kun Qin <kuqin12@gmail.com>