Adding check for underflow/overflow when iterating through resource descriptor hobs (#1549)
During DXE core memory service initialization, the system would check
available resource descriptor hobs against the memory top from PHIT hob.

However, it is possible that a given resource descriptor hob will not be
larger than the cover the memory top, causing the Length calculation to
underflow. Or the hob is not properly prepared and cause the calculation
to overflow.

This change adds a check for potential underflow/overflow before
performing the subtraction.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

This change was tested on QEMU SBSA and booted to both UEFI shell and
Windows VOS.

N/A